+++ features/deactivation/impl/src/test/kotlin/io/element/android/features/logout/impl/AccountDeactivationViewTest.kt
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2024 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ * Please see LICENSE in the repository root for full details.
+ */
+
+package io.element.android.features.logout.impl
+
+import androidx.activity.ComponentActivity
+import androidx.compose.ui.test.junit4.AndroidComposeTestRule
+import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import io.element.android.tests.testutils.EnsureNeverCalled
+import io.element.android.tests.testutils.EventsRecorder
+import io.element.android.tests.testutils.ensureCalledOnce
+import io.element.android.tests.testutils.pressBack
+import org.junit.Rule
+import org.junit.Test
+import org.junit.rules.TestRule
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class AccountDeactivationViewTest {
+    @get:Rule val rule = createAndroidComposeRule<ComponentActivity>()
+
+    @Test
+    fun `clicking on back invokes the expected callback`() {
+        val eventsRecorder = EventsRecorder<AccountDeactivationEvents>(expectEvents = false)
+        ensureCalledOnce {
+            rule.setAccountDeactivationView(
+                state = anAccountDeactivationState(eventSink = eventsRecorder),
+                onBackClick = it,
+            )
+            rule.pressBack()
+        }
+    }
+
+    // TODO Add more tests
+}
+
+private fun <R : TestRule> AndroidComposeTestRule<R, ComponentActivity>.setAccountDeactivationView(
+    state: AccountDeactivationState,
+    onBackClick: () -> Unit = EnsureNeverCalled(),
+) {
+    setContent {
+        AccountDeactivationView(
+            state = state,
+            onBackClick = onBackClick,
+        )
+    }
+}