+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/timeline/components/TimelineItemRow.kt
@@ -30,7 +30,6 @@ import io.element.android.libraries.designsystem.text.toPx
 import io.element.android.libraries.designsystem.theme.highlightedMessageBackgroundColor
 import io.element.android.libraries.matrix.api.core.EventId
 import io.element.android.libraries.matrix.api.core.UserId
-import io.element.android.libraries.matrix.api.timeline.item.event.MessageShield
 
 @Composable
 internal fun TimelineItemRow(
@@ -43,7 +42,6 @@ internal fun TimelineItemRow(
     onLinkClick: (String) -> Unit,
     onClick: (TimelineItem.Event) -> Unit,
     onLongClick: (TimelineItem.Event) -> Unit,
-    onShieldClick: (MessageShield) -> Unit,
     inReplyToClick: (EventId) -> Unit,
     onReactionClick: (key: String, TimelineItem.Event) -> Unit,
     onReactionLongClick: (key: String, TimelineItem.Event) -> Unit,
@@ -115,9 +113,8 @@ internal fun TimelineItemRow(
                             isHighlighted = timelineItem.isEvent(focusedEventId),
                             onClick = { onClick(timelineItem) },
                             onLongClick = { onLongClick(timelineItem) },
-                            onShieldClick = onShieldClick,
-                            onUserDataClick = onUserDataClick,
                             onLinkClick = onLinkClick,
+                            onUserDataClick = onUserDataClick,
                             inReplyToClick = inReplyToClick,
                             onReactionClick = onReactionClick,
                             onReactionLongClick = onReactionLongClick,
@@ -141,7 +138,6 @@ internal fun TimelineItemRow(
                     focusedEventId = focusedEventId,
                     onClick = onClick,
                     onLongClick = onLongClick,
-                    onShieldClick = onShieldClick,
                     inReplyToClick = inReplyToClick,
                     onUserDataClick = onUserDataClick,
                     onLinkClick = onLinkClick,