+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/timeline/focus/FocusRequestStateView.kt
@@ -36,7 +36,7 @@ fun FocusRequestStateView(
             }
             ErrorDialog(
                 content = errorMessage,
-                onDismiss = onClearFocusRequestState,
+                onSubmit = onClearFocusRequestState,
                 modifier = modifier,
             )
         }