+++ features/preferences/impl/src/main/kotlin/io/element/android/features/preferences/impl/PreferencesFlowNode.kt
@@ -20,6 +20,7 @@ import com.bumble.appyx.navmodel.backstack.operation.push
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import io.element.android.anvilannotations.ContributesNode
+import io.element.android.features.deactivation.api.AccountDeactivationEntryPoint
 import io.element.android.features.licenses.api.OpenSourceLicensesEntryPoint
 import io.element.android.features.lockscreen.api.LockScreenEntryPoint
 import io.element.android.features.logout.api.LogoutEntryPoint
@@ -52,6 +53,7 @@ class PreferencesFlowNode @AssistedInject constructor(
     private val notificationTroubleShootEntryPoint: NotificationTroubleShootEntryPoint,
     private val logoutEntryPoint: LogoutEntryPoint,
     private val openSourceLicensesEntryPoint: OpenSourceLicensesEntryPoint,
+    private val accountDeactivationEntryPoint: AccountDeactivationEntryPoint,
 ) : BaseFlowNode<PreferencesFlowNode.NavTarget>(
     backstack = BackStack(
         initialElement = plugins.filterIsInstance<PreferencesEntryPoint.Params>().first().initialElement.toNavTarget(),
@@ -99,6 +101,9 @@ class PreferencesFlowNode @AssistedInject constructor(
 
         @Parcelize
         data object SignOut : NavTarget
+
+        @Parcelize
+        data object AccountDeactivation : NavTarget
 
         @Parcelize
         data object OssLicenses : NavTarget
@@ -150,6 +155,10 @@ class PreferencesFlowNode @AssistedInject constructor(
 
                     override fun onSignOutClick() {
                         backstack.push(NavTarget.SignOut)
+                    }
+
+                    override fun onOpenAccountDeactivation() {
+                        backstack.push(NavTarget.AccountDeactivation)
                     }
                 }
                 createNode<PreferencesRootNode>(buildContext, plugins = listOf(callback))
@@ -235,6 +244,9 @@ class PreferencesFlowNode @AssistedInject constructor(
             }
             is NavTarget.OssLicenses -> {
                 openSourceLicensesEntryPoint.getNode(this, buildContext)
+            }
+            NavTarget.AccountDeactivation -> {
+                accountDeactivationEntryPoint.createNode(this, buildContext)
             }
         }
     }